chore(core): add minimal workshop story
