+++ packages/sanity/src/core/components/commandList/__workshop__/FilterableStory.tsx
@@ -1,4 +1,4 @@
-import {Box, Button, Card, Flex, Stack, Text, TextInput} from '@sanity/ui'
+import {Box, Card, Flex, Stack, Text, TextInput} from '@sanity/ui'
 import {useBoolean} from '@sanity/ui-workshop'
 import React, {ComponentProps, KeyboardEvent, useCallback, useMemo, useState} from 'react'
 import styled from 'styled-components'
@@ -7,7 +7,7 @@ import {CommandListProvider, type CommandListVirtualItemProps} from '../CommandL
 import {CommandListTextInput} from '../CommandListTextInput'
 import {useCommandList} from '../useCommandList'
 
-const ITEMS = [...Array(50000).keys()].map((i) => `Button ${i}`)
+const ITEMS = [...Array(50000).keys()]
 
 const CardContainer = styled(Card)`
   height: 400px;
@@ -15,37 +15,50 @@ const CardContainer = styled(Card)`
   width: 100%;
 `
 
-export default function ButtonStory() {
+const StyledLink = styled.a<{$index: number}>`
+  background: ${({$index}) => ($index % 2 === 0 ? '#1a1a1a' : '#1f1f1f')};
+  display: block;
+  padding: 10px;
+  &[data-active] {
+    background: #333;
+  }
+`
+
+export default function FilterableStory() {
   const [filter, setFilter] = useState<string>('')
   const [message, setMessage] = useState('')
-  const showInput = useBoolean('Show input', false, 'Props')
+  const showInput = useBoolean('Show input', true, 'Props')
 
   const filteredValues = useMemo(() => {
     const values = ITEMS.map((i) => ({value: i}))
     if (!filter) {
       return values
     }
-    return values.filter((i) => i.value.toLowerCase().includes(filter.toLowerCase()))
+    return values.filter((i) => i.value.toString().includes(filter))
   }, [filter])
 
-  const handleChange = useCallback(
-    (e: KeyboardEvent<HTMLInputElement>) => setFilter(e.currentTarget.value),
-    []
-  )
-  const handleChildClick = useCallback((msg: string) => setMessage(msg), [])
-  const handleClear = useCallback(() => setFilter(''), [])
+  const handleChange = useCallback((e: KeyboardEvent<HTMLInputElement>) => {
+    setFilter(e.currentTarget.value)
+  }, [])
+
+  const handleChildClick = useCallback((msg: string) => {
+    setMessage(msg)
+  }, [])
+
+  const handleClear = useCallback(() => {
+    setFilter('')
+  }, [])
 
   const VirtualListItem = useMemo(() => {
-    return function VirtualListItemComponent({value}: CommandListVirtualItemProps<string>) {
-      const handleClick = useCallback(() => handleChildClick(`${value} clicked`), [value])
+    return function VirtualListItemComponent({index, value}: CommandListVirtualItemProps<number>) {
+      const handleClick = useCallback(
+        () => handleChildClick(`Button ${value.toString()} clicked`),
+        [value]
+      )
       return (
-        <Button
-          mode="bleed"
-          onClick={handleClick}
-          style={{borderRadius: 0, width: '100%'}}
-          text={value}
-          tone="primary"
-        />
+        <StyledLink $index={index} onClick={handleClick}>
+          <Text>{value}</Text>
+        </StyledLink>
       )
     }
   }, [handleChildClick])
@@ -54,21 +67,21 @@ export default function ButtonStory() {
     <CardContainer padding={4}>
       <Stack space={3}>
         <CommandListProvider
-          activeItemDataAttr="data-selected"
+          activeItemDataAttr="data-active"
           ariaChildrenLabel="Children"
           ariaInputLabel="Header"
           autoFocus
-          itemComponent={VirtualListItem}
           fixedHeight
+          itemComponent={VirtualListItem}
           values={filteredValues}
           virtualizerOptions={{
-            estimateSize: () => 35,
+            estimateSize: () => 30,
           }}
         >
           <CommandListContent
             filter={filter}
-            onClear={handleClear}
             onChange={handleChange}
+            onClear={handleClear}
             showInput={showInput}
           />
         </CommandListProvider>
@@ -89,7 +102,7 @@ interface CommandListContentProps
 }
 
 const CommandListContent = ({filter, onChange, onClear, showInput}: CommandListContentProps) => {
-  const {values, virtualizer} = useCommandList<string>()
+  const {values, virtualizer} = useCommandList<number>()
 
   /**
    * Scroll command list to the top on filter query change
@@ -103,7 +116,7 @@ const CommandListContent = ({filter, onChange, onClear, showInput}: CommandListC
         }
       }
     },
-    [onChange, values.length, virtualizer]
+    [onChange, virtualizer, values.length]
   )
 
   /**
@@ -121,7 +134,7 @@ const CommandListContent = ({filter, onChange, onClear, showInput}: CommandListC
   return (
     <Flex direction="column" style={{height: '400px'}}>
       {showInput && (
-        <Box marginBottom={2} style={{flexShrink: 0}}>
+        <Box flex="none" marginBottom={2}>
           <CommandListTextInput
             clearButton={filter ? filter.length > 0 : false}
             onClear={handleClear}
